requirements.txt
Flask
flask-bcrypt
mysql-connector-python
